feat: enhance company and user registration response schemas and improve token validation logic
